fjern dobbel a tag
